Found another instance of unsigned int/long that needed to be converted to size_t
git-svn-id: http://svn.atnf.csiro.au/duchamp/trunk@888 450e7745-72da-457f-8a74-b07dc0cda934

+++ src/ATrous/atrous.hh
@@ -51,7 +51,7 @@ namespace duchamp
 			   float *&input,float *&output, Param &par);
 
   /// @brief Subtract a baseline from a set of spectra in a cube. 
-  void baselineSubtract(size_t numSpec, unsigned int specLength, 
+  void baselineSubtract(size_t numSpec, size_t specLength, 
 			float *originalCube, float *baseline, Param &par);
 
   /// @brief Find the baseline of a 1-D spectrum.
